requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
boto3==1.34.104
botocore==1.34.104
celery==5.4.0
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.5.1
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.6
django-filter==24.2
django-simple-history==3.5.0
django-storages==1.14.3
djangorestframework==3.15.1
drf-spectacular==0.27.2
filelock==3.14.0
identify==2.5.36
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.1
pre-commit==3.7.1
prompt-toolkit==3.0.43
psycopg==3.1.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.1
wcwidth==0.2.13
whitenoise==6.6.0
